requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0
nilearn>=0.10.0
nibabel>=5.0.0
matplotlib>=3.7.0
pyyaml>=6.0
umap-learn>=0.5.3
